+++ client/renderer.cpp
@@ -2,29 +2,29 @@
 
 void renderer::gl::draw_outlined_box(BoundingBox box) {
 	glBegin(3);
-	glVertex3f(box.minX, box.minY, box.minZ);
-	glVertex3f(box.maxX, box.minY, box.minZ);
-	glVertex3f(box.maxX, box.minY, box.maxZ);
-	glVertex3f(box.minX, box.minY, box.maxZ);
-	glVertex3f(box.minX, box.minY, box.minZ);
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.minZ));
+	glVertex3f(static_cast<GLfloat>(box.maxX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.minZ));
+	glVertex3f(static_cast<GLfloat>(box.maxX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.maxZ));
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.maxZ));
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.minZ));
 	glEnd();
 
 	glBegin(3);
-	glVertex3f(box.minX, box.maxY, box.minZ);
-	glVertex3f(box.maxX, box.maxY, box.minZ);
-	glVertex3f(box.maxX, box.maxY, box.maxZ);
-	glVertex3f(box.minX, box.maxY, box.maxZ);
-	glVertex3f(box.minX, box.maxY, box.minZ);
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.minZ));
+	glVertex3f(static_cast<GLfloat>(box.maxX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.minZ));
+	glVertex3f(static_cast<GLfloat>(box.maxX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.maxZ));
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.maxZ));
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.minZ));
 	glEnd();
 
 	glBegin(1);
-	glVertex3f(box.minX, box.minY, box.minZ);
-	glVertex3f(box.minX, box.maxY, box.minZ);
-	glVertex3f(box.maxX, box.minY, box.minZ);
-	glVertex3f(box.maxX, box.maxY, box.minZ);
-	glVertex3f(box.maxX, box.minY, box.maxZ);
-	glVertex3f(box.maxX, box.maxY, box.maxZ);
-	glVertex3f(box.minX, box.minY, box.maxZ);
-	glVertex3f(box.minX, box.maxY, box.maxZ);
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.minZ));
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.minZ));
+	glVertex3f(static_cast<GLfloat>(box.maxX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.minZ));
+	glVertex3f(static_cast<GLfloat>(box.maxX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.minZ));
+	glVertex3f(static_cast<GLfloat>(box.maxX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.maxZ));
+	glVertex3f(static_cast<GLfloat>(box.maxX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.maxZ));
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.minY), static_cast<GLfloat>(box.maxZ));
+	glVertex3f(static_cast<GLfloat>(box.minX), static_cast<GLfloat>(box.maxY), static_cast<GLfloat>(box.maxZ));
 	glEnd();
 }
